makes pro_register routes more resourceful, easier to maintain'
