+++ app/controllers/pro_register/category_controller.rb
@@ -1,8 +1,4 @@
 class ProRegister::CategoryController < ApplicationController
-  def index
-    industry = Industry.find(params[:industry_id])
-    @categories = industry.categories.all
-  end
 
   def show
     @category = Category.find(params[:id])
