--- app/controllers/pros_controller/industry_controller.rb
@@ -1,5 +1,0 @@
-class ProRegister::IndustryController < ApplicationController
-  def index
-    @industries = Industry.all
-  end
-end
