+++ app/views/pro/new.html.erb
@@ -9,7 +9,7 @@
     </div>
   </div>
   <div class="form-signin">
-    <%= bootstrap_form_for(@pro, url: pro_register_path, method: :post) do |f| %>
+    <%= bootstrap_form_for(@pro, url: pro_index_path, method: :post) do |f| %>
       <h4>Account Info</h4>
       <%= f.text_field :first_name, placeholder: "First Name", id: "first_name"%>
       <%= f.text_field :last_name, placeholder: "Last Name", id: "last_name"%>
