+++ app/views/pro_register/category/show.html.erb
@@ -1,6 +1,6 @@
 <div class="service-choices">
   <h1><%= @category.name %></h1>
-  <%= form_tag(pro_register_signup_path, method: "get")  do %>
+  <%= form_tag(new_pro_path, method: "get")  do %>
   <table>
     <tbody class="service_table">
       <% @services.each do |service| %>
