+++ app/views/pro_register/industry/show.html.erb
@@ -1,0 +1,5 @@
+<div class="category-choices">
+  <% @industry.categories.each do |category| %>
+    <%= link_to category.name, pro_register_industry_category_path(@industry, category) %><br>
+  <% end %>
+</div>
